requirements.flask.txt
flask
werkzeug
tensorflow
numpy
pillow
